add gui module
