+++ lua/Arduino-Nvim/init.lua
@@ -1,6 +1,7 @@
 local remap = require("Arduino-Nvim.remap")
 local commands = require("Arduino-Nvim.commands")
 local b_config = require("Arduino-Nvim.board_config")
+local gui = require("Arduino-Nvim.gui")
 require("Arduino-Nvim.libGetter")
 local M = {}
 
@@ -8,16 +9,16 @@ b_config.load_or_create_config()
 remap.load_keymaps()
 
 vim.api.nvim_create_user_command("InoSelectBoard", function()
-	commands.select_board_gui()
+	gui.select_board_gui()
 end, {})
 vim.api.nvim_create_user_command("InoSelectPort", function()
-	commands.select_port_gui()
+	gui.select_port_gui()
 end, {})
 vim.api.nvim_create_user_command("InoCheck", function()
 	commands.check()
 end, {})
 vim.api.nvim_create_user_command("InoGUI", function()
-	commands.gui()
+	gui.gui()
 end, {})
 vim.api.nvim_create_user_command("InoMonitor", function()
 	commands.monitor()
@@ -29,7 +30,7 @@ vim.api.nvim_create_user_command("InoUpload", function()
 	commands.upload()
 end, {})
 vim.api.nvim_create_user_command("InoUploadSlow", function()
-	commands.baudrate = "1200"
+	b_config.board_config_table.baudrate = "1200"
 	vim.notify("Trying upload with 1200 baud rate...", vim.log.levels.INFO)
 	commands.upload()
 end, {})
